refactor: folder structure and barrel files
