+++ src/features/navigation/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useActiveSection"
